requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
et_xmlfile==2.0.0
fonttools==4.60.1
frozenlist==1.8.0
idna==3.11
kiwisolver==1.4.9
magic-filter==1.0.12
matplotlib==3.10.7
multidict==6.7.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pipreqs==0.4.13
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarg==0.1.10
yarl==1.22.0
